requirements.txt
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
filelock==3.13.1
fonttools==4.43.1
fsspec==2023.10.0
fvcore==0.1.5.post20221221
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iopath==0.1.10
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
matplotlib==3.8.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.1
numpy-quaternion==2022.4.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
portalocker==2.8.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@2f11ddc5ee7d6bd56f2fb6744a16776fab6536f7
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
scipy==1.11.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
termcolor==2.3.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transforms3d==0.4.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
xkit==0.0.0
yacs==0.1.8
zipp==1.0.0
